pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython==8.10.0",
    "jupyter>=1.0.0,<2",
    "mypy==0.991",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest==7.2.1",
    "pytest-mock==3.11.1",
    "ruff==0.11.11",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7",
    "types-protobuf>=4.24.0.4,<5",
    "types-redis==4.5.5.0",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "black[jupyter]<=23.9.1,>=23.7.0",
    "codespell[toml]>=v2.2.6",
    "diff-cover>=9.2.0",
    "pytest-cov>=6.1.1",
    "numpy>=1.24.0",
]

[project]
name = "llama-index-packs-multimodal-agentic-rag"
version = "0.1.0"
description = "Multimodal Agentic RAG with Sidecar Parsing and Self-Correction Workflow."
authors = [{name = "linsir2", email = "linboweilin20@gmail.com"}]
maintainers = [{name = "linsir2"}]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = "MIT"
keywords = [
    "multimodal",
    "rag",
    "agentic",
    "neo4j",
    "qdrant",
]
dependencies = [
    # ===== Core LlamaIndex =====
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-qdrant",
    "llama-index-graph-stores-neo4j",
    "llama-index-llms-dashscope",
    "llama-index-multi-modal-llms-dashscope",
    "llama-index-embeddings-huggingface",

    # ===== Databases =====
    "qdrant-client>=1.9.0",
    "neo4j>=5.0.0",

    # ===== LLM / API =====
    "dashscope>=1.19.0",
    "httpx>=0.27.0",

    # ===== Gradio Frontend =====
    "gradio>=4.0.0",

    # ===== Optional (if you use streaming or async file handling) =====
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "aiofiles",

    # ===== PDF / Image Parsing =====
    "pdfplumber",
    "pymupdf",
    "pillow",

    # ===== Utilities =====
    "nest-asyncio",
    "python-dotenv",
]


[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.hatch.build.targets.sdist]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.hatch.build.targets.wheel]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.llamahub]
contains_example = false
import_path = "llama_index.packs.multimodal_agentic_rag"

[tool.llamahub.class_authors]
MultimodalAgenticRAGPack = "linsir2"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.10"